Bigger change: Improve handling of dedicated pages
